# Speed-up dagre.Graph