I've noticed that using `Map` instead of `Object` for node's predecessors and successors speeds-up the layouting time.

model | Worker before (ddd5f6f8a32791de7a6d41da6ad5f1537e369aa9) [ms] | Worker after [ms]
------|-----------:|-----------------:
retinanet_resnet50_fpn_v2_Opset18.onnx | 1984 | 1073
yolov9-c.onnx | 1045 | 697
yolov5n.onnx | 184 | 145

(CPU : AMD Ryzen 9 7900X)